Revert "feat: setup `blind-signer` lib and binary"
This reverts commit b04386d4f5950484f1618703ff22b87317b6a138.
Since the signer is ready, this blind signer is not necessary and it is
better to not add this additional component that needs to be kept
updated.
