--- blind-signer/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "blind-signer"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-slog = { version = "2.5.2", features = ["max_level_trace"] }
-pico-args = "0.5.0"
-reqwest = { version = "0.11", default_features = false, features = [
-  "blocking",
-  "json",
-  "rustls",
-  "rustls-tls",
-] }
-serde_json = { version = "1.0", features = [
-  "arbitrary_precision",
-  "raw_value",
-] }
-stacks = { package = "stackslib", path = "../stackslib" }
-stacks-common = { path = "../stacks-common" }
-libsigner = { path = "../libsigner" }
-stacks-node = { path = "../testnet/stacks-node" }
-
-[lib]
-name = "blind_signer"
-path = "src/lib.rs"
-
-[[bin]]
-name = "blind-signer"
-path = "src/main.rs"
