+++ testnet/stacks-node/src/globals.rs
@@ -12,8 +12,8 @@ use stacks::chainstate::stacks::db::StacksChainState;
 use stacks::chainstate::stacks::miner::MinerStatus;
 use stacks::net::NetworkResult;
 use stacks_common::types::chainstate::{BlockHeaderHash, BurnchainHeaderHash, ConsensusHash};
-use stacks_node::config::MinerConfig;
 
+use crate::config::MinerConfig;
 use crate::neon::Counters;
 use crate::neon_node::LeaderKeyRegistrationState;
 use crate::run_loop::RegisteredKey;
