--- testnet/stacks-node/src/lib.rs
@@ -1,9 +1,0 @@
-#[allow(unused_imports)]
-#[macro_use(o, slog_log, slog_trace, slog_debug, slog_info, slog_warn, slog_error)]
-extern crate slog;
-#[macro_use]
-extern crate stacks_common;
-
-pub mod chain_data;
-pub mod config;
-pub mod utils;
