+++ testnet/stacks-node/src/run_loop/neon.rs
@@ -27,7 +27,6 @@ use stacks_common::deps_common::ctrlc::SignalId;
 use stacks_common::types::PublicKey;
 use stacks_common::util::hash::Hash160;
 use stacks_common::util::{get_epoch_time_secs, sleep_ms};
-use stacks_node::config::Config;
 use stx_genesis::GenesisData;
 
 use super::RunLoopCallbacks;
@@ -40,7 +39,9 @@ use crate::node::{
     use_test_genesis_chainstate,
 };
 use crate::syncctl::{PoxSyncWatchdog, PoxSyncWatchdogComms};
-use crate::{run_loop, BitcoinRegtestController, BurnchainController, EventDispatcher, Keychain};
+use crate::{
+    run_loop, BitcoinRegtestController, BurnchainController, Config, EventDispatcher, Keychain,
+};
 
 pub const STDERR: i32 = 2;
