+++ testnet/stacks-node/src/tests/epoch_24.rs
@@ -35,13 +35,12 @@ use stacks_common::types::Address;
 use stacks_common::util::hash::{bytes_to_hex, hex_bytes, Hash160};
 use stacks_common::util::secp256k1::Secp256k1PublicKey;
 use stacks_common::util::sleep_ms;
-use stacks_node::config::{EventKeyType, EventObserverConfig, InitialBalance};
-use stacks_node::utils::{get_account, submit_tx};
 
+use crate::config::{EventKeyType, EventObserverConfig, InitialBalance};
 use crate::tests::bitcoin_regtest::BitcoinCoreController;
 use crate::tests::neon_integrations::{
-    get_chain_info, get_pox_info, neon_integration_test_conf, next_block_and_wait, test_observer,
-    wait_for_runloop,
+    get_account, get_chain_info, get_pox_info, neon_integration_test_conf, next_block_and_wait,
+    submit_tx, test_observer, wait_for_runloop,
 };
 use crate::tests::{make_contract_call, to_addr};
 use crate::{neon, BitcoinRegtestController, BurnchainController};
