+++ testnet/stacks-node/src/tests/stackerdb.rs
@@ -21,14 +21,13 @@ use stacks::chainstate::stacks::StacksPrivateKey;
 use stacks::libstackerdb::{StackerDBChunkAckData, StackerDBChunkData};
 use stacks_common::types::chainstate::StacksAddress;
 use stacks_common::util::hash::Sha512Trunc256Sum;
-use stacks_node::config::{EventKeyType, EventObserverConfig, InitialBalance};
-use stacks_node::utils::submit_tx;
 use {reqwest, serde_json};
 
 use super::bitcoin_regtest::BitcoinCoreController;
 use crate::burnchains::BurnchainController;
+use crate::config::{EventKeyType, EventObserverConfig, InitialBalance};
 use crate::tests::neon_integrations::{
-    neon_integration_test_conf, next_block_and_wait, test_observer, wait_for_runloop,
+    neon_integration_test_conf, next_block_and_wait, submit_tx, test_observer, wait_for_runloop,
 };
 use crate::tests::{make_contract_publish, to_addr};
 use crate::{neon, BitcoinRegtestController};
